requirements.txt
fastapi==0.95.2
pydantic==1.10.13
dbt-core==1.6.0
dbt-snowflake==1.6.0

# Core API dependencies
uvicorn==0.24.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow<19.0.0

# Database and data warehouse
snowflake-connector-python>=3.5.0
snowflake-sqlalchemy>=1.5.0
sqlalchemy>=2.0.0
cryptography>=41.0.0

# Dashboard and visualization
streamlit>=1.28.0
plotly>=5.17.0
altair>=4.2.1,<5.0.0

# Development and utilities
python-dotenv>=1.0.0
setuptools>=68.0.0

# DBT for data transformation
# (already included above)

# Basic utilities
psutil>=5.9.0

# Monitoring
prometheus_client>=0.16.0